+++ CHANGELOG.md
@@ -18,6 +18,10 @@
  - Fix unset channel causing 500 error ([#631](https://github.com/kizniche/mycodo/issues/631))
  - During first install, initialize after install of influxdb
 
+### Features
+
+ - Add Trigger: Infrared Remote Input
+
 ### Miscellaneous
 
  - Add wiringpi to install